+++ apps/explorer/lib/explorer/chain/import/runner/blocks.ex
@@ -168,11 +168,7 @@ defmodule Explorer.Chain.Import.Runner.Blocks do
     query =
       from(
         transaction in where_forked(blocks_changes),
-        select: %{
-          block_hash: transaction.block_hash,
-          index: transaction.index,
-          hash: transaction.hash
-        },
+        select: transaction,
         # Enforce Transaction ShareLocks order (see docs: sharelocks.md)
         order_by: [asc: :hash],
         lock: "FOR UPDATE"
@@ -196,11 +192,7 @@ defmodule Explorer.Chain.Import.Runner.Blocks do
             updated_at: ^updated_at
           ]
         ],
-        select: %{
-          block_hash: s.block_hash,
-          index: s.index,
-          hash: s.hash
-        }
+        select: s
       )
 
     {_num, transactions} = repo.update_all(update_query, [], timeout: timeout)