+++ apps/indexer/lib/indexer/block/fetcher.ex
@@ -13,7 +13,7 @@ defmodule Indexer.Block.Fetcher do
   alias Explorer.Chain
   alias Explorer.Chain.{Address, Block, Hash, Import, Transaction}
   alias Explorer.Chain.Cache.Blocks, as: BlocksCache
-  alias Explorer.Chain.Cache.{Accounts, BlockNumber, Transactions}
+  alias Explorer.Chain.Cache.{Accounts, BlockNumber, PendingTransactions, Transactions}
   alias Indexer.Block.Fetcher.Receipts
 
   alias Indexer.Fetcher.{
@@ -175,7 +175,7 @@ defmodule Indexer.Block.Fetcher do
            ) do
       result = {:ok, %{inserted: inserted, errors: blocks_errors}}
       update_block_cache(inserted[:blocks])
-      update_transactions_cache(inserted[:transactions])
+      update_transactions_cache(inserted[:transactions], inserted[:fork_transactions])
       update_addresses_cache(inserted[:addresses])
       result
     else
@@ -194,8 +194,10 @@ defmodule Indexer.Block.Fetcher do
 
   defp update_block_cache(_), do: :ok
 
-  defp update_transactions_cache(transactions) do
+  defp update_transactions_cache(transactions, forked_transactions) do
     Transactions.update(transactions)
+    PendingTransactions.update_pending(transactions)
+    PendingTransactions.update_pending(forked_transactions)
   end
 
   defp update_addresses_cache(addresses), do: Accounts.drop(addresses)